refactor: remove dead code, harden graph copies, clean warnings
Shallow-copy safety (graph):
- add reference counting to boat_graph_t; boat_graph_copy/subgraph/merge now
  hold a reference to their source graph so the source's node data stays
  alive while the shallow copy exists (was a dangling-pointer hazard).

Dead code removal:
- delete the stale duplicate src/cuda/ directory (CMake only builds cuda/).
- delete the unused boat_computation_graph_* executor (add_operation/
  add_variable were never exposed in the public header, so it was unusable).
- remove genuinely-dead static functions: compute_weight_gradient/_bias
  (conv), compute_forward_{mean,conv,pool,dense} (autodiff), tail_loop_f32
  (simd), sum_last_dim_4d (attention), tensor_shape (swin), get_attr_int
  (onnx), skip_value_info (onnx_pb).

Build hygiene (-Wall -Wextra):
- fix 3 'thread attribute ignored' warnings by using C11 _Thread_local.
- fix unused variables/parameters and sign-compare warnings in memory,
  prune, quantize, activation, arithmetic, linear, prelu, swin,
  transformer_decoder, attention, onnx, onnx_pb, huggingface, graph,
  autodiff.

Note: remaining warnings are const-correctness (-Wdiscarded-qualifiers),
a separate audit; and 5 cJSON-guarded HuggingFace functions that are only
'defined but not used' in the non-cJSON Makefile build (they are live in
the CMake build with BOAT_USE_CJSON).

Tests: 34/34 (Windows mingw64) and 37/37 (WSL2 Linux) pass; full valgrind
memcheck over the Linux suite is 0 errors / 0 bytes leaked.

+++ src/ops/linear.c
@@ -289,7 +289,7 @@ BOAT_API boat_tensor_t* boat_transpose(const boat_tensor_t* a, int dim0, int dim
     if (!a) return NULL;
 
     size_t ndim = boat_tensor_ndim(a);
-    if (dim0 >= ndim || dim1 >= ndim) {
+    if ((size_t)dim0 >= ndim || (size_t)dim1 >= ndim) {
         return NULL;
     }
